fix(firmware): apply embedded code review fixes
- fix(isr): atomically read-clear radio_event_flag with save_and_disable_interrupts() in float_main and surface_main to eliminate ISR-to-main data race
- fix(comms): replace 8-bit XOR checksum with CRC32 (0xEDB88320) in packet_t; change checksum field uint8_t -> uint32_t
- fix(fsm): change ctrl_state field type from uint8_t to BuoyancyControlState_t; replace magic number comparisons (== 1, == 2) with enum constants (CTRL_BRAKING, CTRL_HOVER)
- perf(pid): convert all double -> float in PIDController struct, pid_init, pid_update, pid_reset, pid_set_integral; eliminates slow GCC soft-float double paths on RP2040 (no hardware FPU)
- perf(actuator): convert double -> float in actuator_tick and actuator_vref_set; cache ADC read as cached_pos in tick() to eliminate redundant 32-sample ADC reads per loop; remove dead actuator_move_to() wrapper
- fix(console): change bounds handler sscanf %d -> %u to reject negative values that would silently wrap to large uint16_t
- fix(storage): add __attribute__((packed)) to float_settings_t to eliminate compiler padding in flash; bump SETTINGS_MAGIC 0x20261337 -> 0x20261338 to force clean factory reset; add _Static_assert guard for struct size vs FLASH_PAGE_SIZE
- fix(config): add named constants SHALLOW_BIASED_TARGET_M, SURFACE_DETECTION_M, ADAPTIVE_HOLD_THRESHOLD_M, NEUTRAL_ADC_MIN_VALID, NEUTRAL_ADC_MAX_VALID, NUDGE_WAIT_MS to sw_config.h; replace magic literals throughout float_main.c and float_fsm.c
- fix(hil): move HIL volatile globals hil_depth/hil_pressure to file scope with #ifdef HIL_MODE; remove extern declarations inside function body
- fix(zero): handle_zero now checks ms5837_read() return value and aborts with error message on sensor failure

BREAKING CHANGE: packet_t size changed (uint8_t -> uint32_t checksum); both float and surface firmware must be flashed together. float_settings_t layout changed; first boot will reset flash settings to defaults.

+++ lib/pid/pid.c
@@ -1,30 +1,30 @@
 #include "pid.h"
 #include <stdbool.h>
 
-void pid_init(PIDController *pid, double kp, double ki, double kd, double dt, double output_min, double output_max) {
+void pid_init(PIDController *pid, float kp, float ki, float kd, float dt, float output_min, float output_max) {
     pid->kp = kp;
     pid->ki = ki;
     pid->kd = kd;
     pid->dt = dt;
-    pid->integral = 0.0;
-    pid->prev_error = 0.0;
+    pid->integral = 0.0f;
+    pid->prev_error = 0.0f;
     pid->output_min = output_min;
     pid->output_max = output_max;
     pid->integral_min = output_min;
     pid->integral_max = output_max;
-    pid->integral_gate = 0.0; // Disabled by default
-    pid->prev_D = 0.0;
+    pid->integral_gate = 0.0f; // Disabled by default
+    pid->prev_D = 0.0f;
 }
 
-void pid_update(PIDController *pid, double error, double *output) {
-    double P = pid->kp * error;
+void pid_update(PIDController *pid, float error, float *output) {
+    float P = pid->kp * error;
 
     // One-Way Conditional Integration (Gating)
     // We only gate the integral when we are TOO SHALLOW (descending).
-    // This prevents windup during the long drop, but allows the PID to 
+    // This prevents windup during the long drop, but allows the PID to
     // build maximum buoyancy if we ever go too deep (overshoot).
     bool accumulate = true;
-    if (pid->integral_gate > 0.0) {
+    if (pid->integral_gate > 0.0f) {
         if (error < -pid->integral_gate) { // Too shallow by more than threshold
             accumulate = false;
         }
@@ -36,24 +36,24 @@ void pid_update(PIDController *pid, double error, double *output) {
 
     // Apply integral windup bound (Anti-windup)
     // We limit the integral term's CONTRIBUTION to the specified limits.
-    if (pid->ki != 0) {
-        double i_term_max = pid->integral_max;
-        double i_term_min = pid->integral_min;
+    if (pid->ki != 0.0f) {
+        float i_term_max = pid->integral_max;
+        float i_term_min = pid->integral_min;
 
         if (pid->ki * pid->integral > i_term_max) pid->integral = i_term_max / pid->ki;
         if (pid->ki * pid->integral < i_term_min) pid->integral = i_term_min / pid->ki;
     }
 
-    double I = pid->ki * pid->integral;
-    
+    float I = pid->ki * pid->integral;
+
     // Derivative calculation with EMA Filtering
     // Alpha (0.3) weights the newest measurement; (0.7) weights the historical state.
     // This damps rapid fluctuations from sensor noise/bubbles.
-    double raw_D = pid->kd * ((error - pid->prev_error) / pid->dt);
-    double D = (0.3 * raw_D) + (0.7 * pid->prev_D);
+    float raw_D = pid->kd * ((error - pid->prev_error) / pid->dt);
+    float D = (0.3f * raw_D) + (0.7f * pid->prev_D);
     pid->prev_D = D;
 
-    double raw_output = P + I + D;
+    float raw_output = P + I + D;
 
     // Clamp total output to limits
     if (raw_output > pid->output_max) raw_output = pid->output_max;
@@ -64,18 +64,18 @@ void pid_update(PIDController *pid, double error, double *output) {
 }
 
 void pid_reset(PIDController *pid) {
-    pid->integral = 0.0;
-    pid->prev_error = 0.0;
-    pid->prev_D = 0.0;
+    pid->integral = 0.0f;
+    pid->prev_error = 0.0f;
+    pid->prev_D = 0.0f;
 }
 
-void pid_set_integral(PIDController *pid, double integral_value) {
-    // We allow setting the integral directly. 
+void pid_set_integral(PIDController *pid, float integral_value) {
+    // We allow setting the integral directly.
     // This is useful for "seeding" the PID with a baseline guess (like Neutral ADC)
     // and for anti-windup checks.
-    if (pid->ki != 0) {
+    if (pid->ki != 0.0f) {
         pid->integral = integral_value / pid->ki;
     } else {
-        pid->integral = 0;
+        pid->integral = 0.0f;
     }
 }
