+++ lib/pid/pid.h
@@ -2,23 +2,23 @@
 #define PID_H
 
 typedef struct {
-    double kp;
-    double ki;
-    double kd;
-    double dt;
-    double integral;
-    double prev_error;
-    double output_min;
-    double output_max;
-    double integral_min;  // Minimum value for the integral contribution
-    double integral_max;  // Maximum value for the integral contribution
-    double integral_gate; // Conditional integration threshold (0 to disable)
-    double prev_D;        // Previous filtered derivative state for EMA
+    float kp;
+    float ki;
+    float kd;
+    float dt;
+    float integral;
+    float prev_error;
+    float output_min;
+    float output_max;
+    float integral_min;  // Minimum value for the integral contribution
+    float integral_max;  // Maximum value for the integral contribution
+    float integral_gate; // Conditional integration threshold (0 to disable)
+    float prev_D;        // Previous filtered derivative state for EMA
 } PIDController;
 
-void pid_init(PIDController *pid, double kp, double ki, double kd, double dt, double output_min, double output_max);
-void pid_update(PIDController *pid, double error, double *output);
+void pid_init(PIDController *pid, float kp, float ki, float kd, float dt, float output_min, float output_max);
+void pid_update(PIDController *pid, float error, float *output);
 void pid_reset(PIDController *pid);
-void pid_set_integral(PIDController *pid, double integral_value);
+void pid_set_integral(PIDController *pid, float integral_value);
 
 #endif
